requirements.txt
rich>=13.0.0
click>=8.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0